third_party/blink/web_tests/platform/generic/http/tests/devtools/command-line-api-inspect-expected.txt

Tests that inspect() command line api works.


Running: testRevealElement


WebInspector.inspect called with: p#p1
WebInspector.inspect's hints are: []
inspect($('#p1')) = 
Selected node id: 'p1'.
